requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# AI Integration
anthropic>=0.40.0
openai>=1.0.0

# Data Analysis
scikit-learn>=1.3.0
scipy>=1.11.0

# Reporting
jinja2>=3.1.0
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
